refactor: replace logic for linear theme
